refactor MacFileBasedLock and LinuxFileBasedLock
